no chnges of models of non-solutions

+++ runtime/src/main/java/org/modelingvalue/dclare/mps/DModule.java
@@ -56,30 +56,31 @@ public class DModule extends DFromOriginalObject<SModule> implements SModule {
 
     private static final Constant<Pair<Boolean, Set<SLanguage>>, DModuleType> MODULE_TYPE    = Constant.of("MODULE_TYPE", DModuleType::new);
 
-    protected static final DObserved<DModule, Set<DModel>>                    MODELS         = DObserved.of("MODELS", Set.of(), (dModule, pre, post) -> {
-                                                                                                 Set<DModel> ist = models(dModule.original()).sequential().map(DModel::of).toSet();
-                                                                                                 if (!ist.equals(post)) {
-                                                                                                     Setable.<Set<DModel>, DModel> diff(ist, post,                                                  //
-                                                                                                             DMatchedObject::original,                                                              //
-                                                                                                             r -> new ModelDeleteHelper(r.tryOriginal()).delete());
-                                                                                                     return true;
-                                                                                                 } else {
-                                                                                                     return false;
+    protected static final DObserved<DModule, Set<DModel>>                    MODELS         = DObserved.of("MODELS", Set.of(), (m, pre, post) -> {
+                                                                                                 if (m.isSolution()) {
+                                                                                                     Set<DModel> ist = m.models().sequential().map(DModel::of).toSet();
+                                                                                                     if (!ist.equals(post)) {
+                                                                                                         Setable.<Set<DModel>, DModel> diff(ist, post,                                    //
+                                                                                                                 DMatchedObject::original,                                                //
+                                                                                                                 r -> new ModelDeleteHelper(r.tryOriginal()).delete());
+                                                                                                         return true;
+                                                                                                     }
                                                                                                  }
+                                                                                                 return false;
                                                                                              }, containment);
 
     protected static final Observed<DModule, Set<SLanguage>>                  LANGUAGES      = Observed.of("LANGUAGES", Set.of(), (tx, o, pre, post) -> {
-                                                                                                 Setable.<Set<SLanguage>, SLanguage> diff(pre, post,                                                //
-                                                                                                         a -> DClareMPS.ALL_LANGUAGES.set(dClareMPS(), Set::add, a),                                //
+                                                                                                 Setable.<Set<SLanguage>, SLanguage> diff(pre, post,                                      //
+                                                                                                         a -> DClareMPS.ALL_LANGUAGES.set(dClareMPS(), Set::add, a),                      //
                                                                                                          r -> {
                                                                                                          });
                                                                                              }, synthetic);
 
-    private static final Observer<DModule>                                    LANGUAGES_RULE = DObject.observer(LANGUAGES, o -> dClareMPS().read(() -> languages(o.original()))                     //
+    private static final Observer<DModule>                                    LANGUAGES_RULE = DObject.observer(LANGUAGES, o -> dClareMPS().read(() -> languages(o.original()))           //
             .addAll(MODELS.get(o).flatMap(DModel::allUsedLanguages)));
 
     private static final Action<DModule>                                      READ_MODELS    = Action.of("$READ_MODELS", m -> {
-                                                                                                 MODELS.set(m, dClareMPS().read(() -> models(m.original()).sequential().map(DModel::read).toSet()));
+                                                                                                 MODELS.set(m, dClareMPS().read(() -> m.models().sequential().map(DModel::read).toSet()));
                                                                                              }, Priority.urgent);
     @SuppressWarnings("rawtypes")
     protected static final Set<Observer>                                      OBSERVERS      = DObject.OBSERVERS.add(LANGUAGES_RULE);
@@ -242,9 +243,9 @@ public class DModule extends DFromOriginalObject<SModule> implements SModule {
         return Collection.of(module.getUsedLanguages()).sequential().toSet();
     }
 
-    protected static Set<SModel> models(SModule module) {
+    protected Set<SModel> models() {
         Set<SModel> ist = Set.of();
-        for (SModel child : module instanceof Language ? ((Language) module).getAccessoryModels() : module instanceof Solution ? module.getModels() : Set.<SModel> of()) {
+        for (SModel child : isLanguage() ? ((Language) original()).getAccessoryModels() : isSolution() ? original().getModels() : Set.<SModel> of()) {
             if (child instanceof EditableSModel) {
                 ist = ist.add(child);
             }
